After-merge cleanup and preparations migrations and data cleanup tasks
Former-commit-id: d8087e80710bb6d67ac744806cac27af40603964 [formerly b70e13ceaceb8ab5a0c584a9fe14e4725098dc78] [formerly 860d084c9d015b209240b82fb190e59efe137700 [formerly 605deb27156ddccea6d724e86b6523c1d64f7ff9]] [formerly b0991504f998368457e8a51f9e72f7930dcf3a09 [formerly b52e63e1f0bd0396cd88148666319b4c844429cf] [formerly eed2bb9e7c446a34ef80b20bd55bcc96494d1d97 [formerly b4f85dc169bab1ae2faff481581b5a9cb6e00a59]]]
Former-commit-id: ad90051c769ab0d88c9711b3ea43b999539bfdec [formerly 69b6aaebd81d1f987cd03788de51d32a0a97dfaa] [formerly 237d8a2d70cc0a765121561e3e87e37c211ab617 [formerly 8dd18c1267ae009f0a8b145a1659492aea7ede85]]
Former-commit-id: f76327931e4314b4a0645e4151edfa6167b0bfe4 [formerly 0ee1943c951e000298ace77be51259c14c11f961]
Former-commit-id: e9d6a726bac659db76b706a51f50fb8c97c651bc

+++ db/migrate/20150514120739_add_primary_key_to_account_organizations.rb
@@ -1,9 +1,13 @@
 class AddPrimaryKeyToAccountOrganizations < ActiveRecord::Migration
   def up
-    add_column :accounts_organizations, :id, :primary_key
+    if Apartment::Tenant.current == 'public' && !column_exists?(:accounts_organizations, :id)
+      add_column :accounts_organizations, :id, :primary_key
+    end
   end
 
   def down
-    remove_column :accounts_organizations, :id
+    if Apartment::Tenant.current == 'public' && column_exists?(:accounts_organizations, :id)
+      remove_column :accounts_organizations, :id
+    end
   end
 end
